Compress gimp/main-gimp.c, and convert help from http to pop-up message.
Although TABs were used often with gimp2, noticed that spaces are chosen
for gimp3 therefore converted code to using spaces and tried maintaining
some consistencies on formatting. Opted for using the more compact style
where you can see more source code in the display, instead of the coding
style which uses a lot of lines and less to see in the display screen.

doc was unfinished, therefore nothing really to http: to. Converted help
to use built-in pop-up message instead. Added extra info in plugin Help.