Partial updates and restructure from GIMP-2.10 to GIMP-3.0
Many changes made to get this from GIMP-2.10 to GIMP-3.0, but at this
point, still does not compile as more edits and fixes are required to
make this successfully compile under GIMP-3.x.

This patch was made as a visible update concerning issue #1 as to the
current status of gimp-refocus-it. More info can be found in issue #1